Moved and renamed references to project model
