+++ api/ai/new/multiple_round_robin_with_adjusted_winner/utils.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from api.models.project import ProjectRequirement
+from api.models.project.project import ProjectRequirement
 from api.models.student import Student
